lint: clear the 5 clang-tidy bugprone errors blocking `make lint`
`make lint` was red on HEAD (5 bugprone-* warnings promoted to errors by
.clang-tidy WarningsAsErrors), which also blocks `make all`. Pure cleanup:
no behaviour change, no IFR logic touched.

- engine.cpp openair_alt + atc_ui.cpp draw_ifr_tab (the copy of it):
  `transition_alt_ft` is an int compared against 0.0f and fed to a float
  ternary -> 4x bugprone-narrowing-conversions. Compare as int, convert
  explicitly. Identical arithmetic.

- openair_db.cpp load_file: bugprone-branch-clone on the DELEGATED/SKYGUIDE/
  MUAC branch, which assigns the same AirspaceClass::CTA as the "CTA" branch.
  SUPPRESSED (NOLINTBEGIN/END), not merged: the branches must stay separate
  and in this order, because a delegation polygon whose name also carries a
  type word ("DELEGATED BY LIMM TO LJLA FIR/UIR") has to be classified by
  that word -- the FIR test has to run first. Merging would silently
  reclassify such overlays as CTA. Rationale recorded in a comment.

Also folded in, same files, not error-level but flagged by the same run:

- airport_overrides.{hpp,cpp}: init(std::string) -> init(const std::string &)
  (performance-unnecessary-value-param; both call sites pass an lvalue), and
  `const std::string key = it.key()` -> const ref (it.key() already returns
  one).

- xplane_context_runtime.cpp select_active_runway: guard the calm-wind branch
  with `if (!best) return "";` before dereferencing. Unreachable today
  (runways.empty() returns above, so the loop always assigns on its first
  pass) but the analyser cannot prove it through the container, and the deref
  would be fatal if that invariant ever changed.

Verified: make lint EXIT=0, 0 errors. make build OK. make test 41 scenarios /
222 assertions PASS, test-stars ALL PASS, test-afis ALL PASS.

+++ src/atc/engine.cpp
@@ -317,7 +317,9 @@ static float approach_gate_ceiling_ft(const xplane_context::XPlaneContext &ctx) 
 // standard-day flights are byte-identical; only non-standard QNH changes (and was
 // wrong before). *** Only for openair_db; airspace_db (atc.dat) is untouched.
 static int openair_alt(const xplane_context::XPlaneContext &ctx) {
-  const float ta = ctx.transition_alt_ft > 0.0f ? ctx.transition_alt_ft : 5000.0f;
+  const float ta = ctx.transition_alt_ft > 0
+                       ? static_cast<float>(ctx.transition_alt_ft)
+                       : 5000.0f;
   return static_cast<int>(ctx.altitude_ft_msl > ta ? ctx.pressure_alt_ft
                                                     : ctx.altitude_ft_msl);
 }
